Refactor xcm tests
